Simple_FastApi_blog/backend/requirements.txt
fastapi==0.115.2
passlib==1.7.4
pydantic==1.10.18
python-dotenv==1.0.1
python-slugify==8.0.4
python_jose==3.3.0
SQLAlchemy==2.0.36
starlette==0.41.0
